improvement(store): moved and updated store's logic
revamp of whole playback logic, replaced 'viewShit' action in vuex store as the playback-to-UI
control engine with HTMLAudioElement listeners, changed some variable names, removed demo playlist
from store adding it programmatically in /Player.vue

BREAKING CHANGE: UI is updated by listening to HTMLAudioElement events instead of updating it
manually in 'viewShit'. This fixes all of the currently known bugs that were as a result of the
previous implementation.

fix #10